4.21 bloglist expansion, step9

+++ part4/controllers/blogs.js
@@ -38,8 +38,26 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).end()
+  }
+  const blogCreatorId = blog.user.toString()
+
+  const token = request.token
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blogDeletorId = user._id.toString()
+
+  if (blogCreatorId === blogDeletorId) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).end()
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
